scripts/requirements.txt
numpy
soundfile
pyserial